Control de propiedad a métodos del API

+++ controllers/api/apiProductsController.js
@@ -1,10 +1,14 @@
 import Product from "../../models/Product.js"
 import {unlink} from 'node:fs/promises'
 import path from "node:path"
+import createError from 'http-errors'
+import { log } from "node:console"
 
 export async function list(req, res, next) {
 
   try {
+
+    const userId = req.apiUserId
 
     const filterName = req.query.name
     const filterTags = req.query.tags
@@ -15,7 +19,7 @@ export async function list(req, res, next) {
     const withCount = req.query.count === 'true'
 
     const filter = {
-            //TODO
+            owner: userId
         }
 
     if (filterName) {
@@ -43,7 +47,9 @@ export async function list(req, res, next) {
 export async function getOne(req,res,next) {
   try {
     const productId = req.params.productId
-    const product = await Product.findById(productId)
+    const userId = req.apiUserId
+
+    const product = await Product.findOne({ _id: productId, owner: userId })
 
     res.json({ result: product })
 
@@ -55,8 +61,12 @@ export async function getOne(req,res,next) {
 export async function newProduct(req, res, next) {
   try {
     const productData = req.body
+    const userId = req.apiUserId
+
     const product = new Product(productData)
     product.image = req.file?.filename
+    product.owner = userId
+
     const savedProduct = await product.save()
 
     res.status(201).json({ result: savedProduct })
@@ -69,10 +79,15 @@ export async function newProduct(req, res, next) {
 export async function update(req, res, next) {
   try {
     const productId = req.params.productId
+    const userId = req.apiUserId
+
     const productData = req.body
     productData.image = req.file?.filename
 
-    const updatedProduct = await Product.findByIdAndUpdate(productId, productData, {
+    const updatedProduct = await Product.findOneAndUpdate({
+      _id: productId,
+      owner: userId
+    }, productData, {
       new: true
     })
 
@@ -86,8 +101,20 @@ export async function update(req, res, next) {
 export async function deleteProduct(req, res, next) {
   try {
     const productId = req.params.productId
+    const userId = req.apiUserId
 
     const product = await Product.findById(productId)
+
+    if (!product) {
+      console.log(`WARNING! user ${userId} trying to delete non existing product`)
+      return next(createError(404))
+    }
+
+    if (product.owner.toString() !== userId) {
+      console.log(`WARNING! user ${userId} trying to delete products from other users`);
+      return next(createError(401))
+    }
+
     if (product.image) {
       await unlink(path.join(import.meta.dirname, '..', '..', 'public', 'productimg', product.image))
     }
